requirements.txt

Flask
python-dotenv
requests
Pyrogram
TGCrypto

Pillow
Flask-SQLAlchemy
PyMySQL
